refactor: common -> response 패지명 변경
